Fix response to resemble sync endpoints
Fix issue where the response returned to client does not resemble the
response typically returned to clients by -sync endpoint counterparts.
The ProxyRequestService must resemble the structure of the -sync
counterparts with respect to where the jobId is found, at a minimum.
This change moves the `jobId` and `itemBarcode` properties from the root
level of the response to inside a new `data` property, which is the
convention for the endpoints currently proxied.